requirements.txt
kivy==2.2.1
pytesseract==0.3.10
Pillow==9.5.0
numpy==1.24.3
pytest==7.4.0
python-dotenv==1.0.0
SQLAlchemy==2.0.20
reportlab==4.0.4  # Para generar PDFs 
opencv-python==4.8.0.74 # Agregar OpenCV 